pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "conversation-fixture"
version = "0.1.0"
description = "Multi-turn conversation test fixture manager for LLMs"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Cognition Commons", email = "tools@cognitioncommons.dev"}
]
keywords = [
    "llm",
    "testing",
    "conversation",
    "fixtures",
    "validation",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Testing",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "httpx>=0.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
    "types-PyYAML",
]

[project.scripts]
conversation-fixture = "conversation_fixture.cli:main"

[project.urls]
Homepage = "https://github.com/cognitioncommons/conversation-fixture"
Repository = "https://github.com/cognitioncommons/conversation-fixture"
Documentation = "https://github.com/cognitioncommons/conversation-fixture#readme"

[tool.setuptools.packages.find]
include = ["conversation_fixture*"]

[tool.black]
line-length = 88
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --cov=conversation_fixture --cov-report=term-missing"
